Get correct name for F# Interactive on *nix
The standard name of F# interactive on unix is "fsharpi"  (fsharpc is
the compiler, not fsc since that clashes with another tool. fsharpi is
named for consistency with fsharpc).

Some older F# 2.0 installs on unix use fsi.

This probes for fsharpi.

To test, I built FAKE (on Mac using F# 3.0's new xbuild)

   xbuild FAKE.sln

It built!!  And then ran FAKE.exe to try to build FAKE with the new FAKE

   mono tools/FAKE/FAKE.exe

We still get an error:

/Users/funnelweb/Projects/GitHub/funnelweb/FAKE/build.fsx(1,1): warning
FS0211: The search directory
'/Users/funnelweb/Projects/GitHub/funnelweb/FAKE/tools\FAKE' could not
be found

That looks like a path separator problem '\', anyway, we're a little
way to getting off the ground here….