Fixed issue #49, Fake.Deploy failures.
FSI didn't seem to like executing the scripts from the system32 folder
despite LocalSystem having full access to this folder. I have added a
workDirectory applicationSetting to the Fake.Deploy agent and it now
appears to fix the problem when it is set to a path outside of System32.
Additionally I have made detecting FSI when only Fsharp 3.0 is installed
possible, as well as making the detection of the ProgramFile based on
CPU architecture more robust.