Fix: Do not append a newline if using ReplaceInFile:
Use File.WriteAllText in ReplaceFile instead of WriteStringToFile false.
WriteStringToFile false has the behaviour, that in every case a newline is appended, because the text is passed as a singleton list of lines and WriteToFile appends each line using a StreamWriter.WriteLine thus resulting in the last line introducing an additional newline.