NuGet: support building symbol+source packages without project file
NuGet supports building symbol/source packages automatically when
specifying the project file. However, it is also possible to build
such packages manually without a project file but instead a nuspec file,
where the symbols and sources are specified explicitly.

The default behavior is unchanged: symbol packages are built if and only
if a project file is provided (in this case we generate, as before, first
both packages with the project file, then overwrite the normal package
using the generated nuspec). If symbol packages should be built with the
nuspec file, then specify p.Symbols = NugetSymbols.Nuspec.