Add support for FxCop /ignoregeneratedcode parameter
To ignore generated code when running FxCop, we can now specify the
parameter

IgnoreGeneratedCode = true

in the parameters for the FxCop task.  The default for this parameter is
false,
in order to provide backwards compatibility with previous versions of
FAKE.

`Generated code` is code which has been decorated with the
`System.CodeDom.Compiler.GeneratedCode` attribute, for example, code
generated
when using the Add Service Reference wizard in Visual Studio.