Change RunService / RunRemoteService to block.
Although the SC command does not wait for the service to finish
executing the service command, we want to at least allow the SC
process to exit so that FAKE doesn't try to kill it. I also changed
it to throw an exception if SC returns a nonzero exist status.