NuGet: support fallback framework groups
Since NuGet 2.0, dependencies and references can be specified by target
framework, where only the entries from the matching group are considered.
NuGet also allows defining a fallback group where the targetFramework is
missing, which is used if none of the other groups match.

Previously FAKE did emit the empty targetFramework attribute, which causes
NuGet to complain. This change skipts the attribute if its value is null
or empty, in compliance with the nuget specs.