Adding ability to control type library export (/tlb flag) of RegAsm.
The previous implementation of RegAsm always exports a type library (by using the /tlb flag). We had a use case, were only registration of the assembly was possible, but no export of a type library (C++ COM+ wrapper).
This fix allows to specify whether to run RegAsm with type library export or not (default: with type library export).

Also escaping assembly paths in command line calls.