Get package version from nuspec file
I was using `GetPackageVersion` in FsLab. When we use Paket to get FsLab dependencies, it
downloads the package `Google.DataTable.Net.Wrapper` and puts it in a file called
`Google.DataTable.Net.Wrapper.3.1.2.nupkg`. However, the `nuspec` file says that the version
is actually `3.1.2.0` and if you add the reference using NuGet, it places it in a directory
with `3.1.2.0` in the name (and also includes the additional `.0` in the  `nupkg` file name).

I'm not sure if this is a bug in Paket (uses wrong file name) or if it is just NuGet giving
us inconsistent information. I suspect the latter, so I'm sending a PR. But perhaps this is
better handled in Paket.. not sure!