Function to remove Compile elems missing files
As well as keeping project files in sync, I also need to be able to remove compile nodes from a project, where the Include attribute points to a file that is missing.  This is useful when I have two projects including the same files, and I delete a bunch of them from the 'master' project I happen to be working on.

This PR adds two functions.  RemoveCompileNodesWithMissingFiles is intended for build script use and removeCompileNodesWithMissingFiles exists for testing purposes, because unlike the former, it does not write to the file system.  I noticed this convention elsewhere in FAKE, but I'm not sure if it is considered idiomatic.  Let me know.