Warn when fail to parse cmd line args. Close #1082
Print a warning when fail to parse command line args with Argu.  We
still fall back to try and run with the pre-2.1.8 argument style.

Extracted exception formatting from the traceException function, so it
could be used when tracing an exception as a warning.

This should help avoid confusion, as witnessed in #1072 and #594.