Add initial support for Fuchu tests
Something like this has been discussed at #234, and there is an open
issue for this more specifically at #742. And while there seem to be
some questions of how to do this in the best way, or polish it, I am of
the opinion that it is better to have something than nothing. It is
rather reassuring to n00bs particularly when they see that tools are at
least aware of each other, and this is one of the big draws of FAKE.

The code is derived from the example from Fuchu itself:

    https://github.com/mausch/Fuchu/blob/5ef9ec303abf56f8ca5511cda3fb53bfa705007c/build.fsx#L119