Change how npm.cmd is located in NpmHelper
I've noticed that Yarn adds `nodejs\bin` to the path variable, which would
return in the old method used to find npm.cmd, but not actually contain
the npm.cmd file. I also removed the fallback to the npm.js package, as it
is silly to assume where the package directory is. If people are using
that package they should overwrite the NpmFilePath variable.