Refactor to use timeout per Octokit API
The issue referred in the code https://github.com/octokit/octokit.net/issues/963
has since been fixed.

The timeout can be set via a method on the client, and per the last
comment https://github.com/octokit/octokit.net/issues/963#issuecomment-399592938
this needs to be set on the release upload request as well. If not set,
the default per-request timeout of 100 seconds is configured.