Handle multiple packages in windows command line
Instead of generating a string that contains multiple package arguments and passing this to the command line generate (which on windows was enclosing all these in quotes), this change appends a list of all the indivitual package arguments to the list of the other arguments.

This stops the args.ToWindowsCommandLine function from inserting quotes around what should be seperate parameters.